requirements.txt
anthropic>=0.40.0
volcengine-python-sdk>=1.0.0
volcengine>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
tos>=2.8.0
Pillow>=10.0.0
itsdangerous>=2.2.0
faster-whisper>=1.1.0
